+++ romancal/resample/resample.py
@@ -295,6 +295,7 @@ class ResampleData:
         )
 
         log.info("Resampling science data")
+        members = []
         for img in self.input_models:
             inwht = resample_utils.build_driz_weight(
                 img,
@@ -323,7 +324,14 @@ class ResampleData:
                 ymax=ymax,
             )
             del data, inwht
-            output_model.meta.resample.members.append(str(img.meta.filename))
+            members.append(str(img.meta.filename))
+
+        members = (
+            members
+            if self.input_models.filepaths is None
+            else self.input_models.filepaths
+        )
+        output_model.meta.resample.members = members
 
         # Resample variances array in self.input_models to output_model
         self.resample_variance_array("var_rnoise", output_model)